move cmd into bramble
