+++ pkg/bramble/cmd-examples.bramble
@@ -1,5 +1,5 @@
 """
-Bramblescript examples
+Bramble cmd examples
 """
