+++ pkg/bramble/integration_test.go
@@ -1,6 +1,7 @@
 package bramble
 
 import (
+	"fmt"
 	"io/ioutil"
 	"os"
 	"testing"
@@ -20,6 +21,22 @@ func TestIntegration(t *testing.T) {
 	// set a unique bramble store for these tests
 	os.Setenv("BRAMBLE_PATH", dir)
 	if err := b.test([]string{"../../tests"}); err != nil {
+		fmt.Printf("%+v", err)
+		t.Error(err)
+	}
+
+	_ = os.RemoveAll(dir)
+}
+
+func TestRunStarlarkBuilder(t *testing.T) {
+	t.Skip()
+	b := Bramble{}
+
+	dir, err := ioutil.TempDir("", TestTmpDirPrefix)
+	assert.NoError(t, err)
+	// set a unique bramble store for these tests
+	os.Setenv("BRAMBLE_PATH", dir)
+	if err := b.run([]string{"../../tests/starlark-builder:run_busybox"}); err != nil {
 		t.Error(err)
 	}
