--- pkg/bramblecmd/README.md
@@ -1,7 +1,0 @@
-
-
-
-- command starts to run
-- stdout and stderr are written to a buffer with a limit, stout/err writes will block when the buffer is full
-- user can pass print_output option if they want to see the output
--
