--- pkg/bramblecmd/buffered_pipe.go
@@ -1,73 +1,0 @@
-package bramblecmd
-
-import (
-	"io"
-	"sync"
-	"syscall"
-)
-
-// https://github.com/golang/go/blob/0436b162397018c45068b47ca1b5924a3eafdee0/src/net/net_fake.go#L173
-
-func newBufferedPipe(softLimit int) *bufferedPipe {
-	p := &bufferedPipe{softLimit: softLimit}
-	p.rCond.L = &p.mu
-	p.wCond.L = &p.mu
-	return p
-}
-
-type bufferedPipe struct {
-	softLimit int
-	mu        sync.Mutex
-	buf       []byte
-	closed    bool
-	rCond     sync.Cond
-	wCond     sync.Cond
-}
-
-func (p *bufferedPipe) Read(b []byte) (int, error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	for {
-		if p.closed && len(p.buf) == 0 {
-			return 0, io.EOF
-		}
-		if len(p.buf) > 0 {
-			break
-		}
-		p.rCond.Wait()
-	}
-
-	n := copy(b, p.buf)
-	p.buf = p.buf[n:]
-	p.wCond.Broadcast()
-	return n, nil
-}
-
-func (p *bufferedPipe) Write(b []byte) (int, error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	for {
-		if p.closed {
-			return 0, syscall.ENOTCONN
-		}
-		if len(p.buf) <= p.softLimit {
-			break
-		}
-		p.wCond.Wait()
-	}
-
-	p.buf = append(p.buf, b...)
-	p.rCond.Broadcast()
-	return len(b), nil
-}
-
-func (p *bufferedPipe) Close() (err error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	p.closed = true
-	p.rCond.Broadcast()
-	p.wCond.Broadcast()
-	return nil
-}
