--- pkg/bramblecmd/examples/imports_scratch.bramble
@@ -1,12 +1,0 @@
-"""
-fake code that tries to reason through how imports work
-"""
-
-load("nomad")
-
-# loads a struct with all global methods
-
-
-def main(args):
-    cmd.cd("..")
-    nomad.deploy()
