--- pkg/bramblecmd/examples/new_structure/main_test.bramble
@@ -1,5 +1,0 @@
-load("github.com/maxmcd/bramble/pkg/bramblecmd/examples/new_structure/main", "ls")
-
-def test_ls():
-    ran = ls()
-    assert.eq(ran.exit_code, 0)
