--- pkg/bramblecmd/examples/other_new_structure/main.bramble
@@ -1,5 +1,0 @@
-load("github.com/maxmcd/bramble/pkg/bramblecmd/examples/new_structure/main", foo="ls")
-
-
-def main():
-    foo()
