--- pkg/bramblecmd/refactor.go
@@ -1,1 +1,0 @@
-package bramblecmd
