--- pkg/bramblecmd/refactor_test.go
@@ -1,1 +1,0 @@
-package bramblecmd
