+++ tests/starlark-builder.bramble
@@ -1,0 +1,49 @@
+def run_busybox():
+    bb = busybox()
+    # whoami is now available to run
+    bb.cmd("whoami").print()
+
+
+def busybox():
+    # download the executable, this is the only way you are allowed
+    # to use the network during builds
+    download = derivation(
+        name="busybox_download",
+        builder="fetch_url",
+        env={
+            "url": "https://brmbl.s3.amazonaws.com/busybox-x86_64.tar.gz",
+            "hash": "2ae410370b8e9113968ffa6e52f38eea7f17df5f436bd6a69cc41c6ca01541a1",
+        },
+    )
+
+    return derivation(name="busybox", builder=build_busybox, build_inputs=[download])
+
+
+def build_busybox():
+    os.mkdir("$out/bin")  # using the builtin os module
+    print("hi")
+    # move the busybox executable into the output for this build
+    cmd(
+        "$busybox_download/busybox-x86_64",
+        "cp",
+        "$busybox_download/busybox-x86_64",
+        "$out/bin/busybox",
+    ).wait()
+    commands = []
+    commands_started = False
+
+    # run the busybox help text to extract supported commands
+    for line in cmd("$out/bin/busybox").output:
+        if commands_started:
+            for name in line.strip().split(","):
+                commands.push(name.strip())
+        if "Currently defined functions" in line:  # the builtin os module
+            commands_started = True
+
+    # move into the output directory so that symlinks are relative
+    os.cd("$out/bin")
+
+    cmd("./busybox ln -s busybox ln").wait()
+    # link each command
+    for cmd in commands:
+        cmd("./ln -s busybox %s" % cmd).wait()
